requirements_phase1.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
pyyaml==6.0.1
tqdm==4.66.1